Add a new config option "LowerThirdHeight", which allows you to specify (in pixels) the height of the lower-third box.
Text will be anchored to the centre of this space. If you don't specify this value, the old behaviour of anchoring to the bottom of the screen still applies.
I've also moved the text update code out to a timed-function, so we can continue to work on the display refresh glitches.